Updated C# domain models to account for Python response changes

+++ src/backend/MyAIRunningMate.Python/app/domain/models/ingestion/user_metrics.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 
 class UserMetrics(BaseModel):
     user_volumetric_oxygen_max: Optional[float] = None
-    user_max_heart_rate: Optional[float] = None
+    user_max_heart_rate: Optional[int] = None
     user_lactate_threshold_heart_rate: Optional[float] = None
     user_lactate_threshold_power: Optional[float] = None
     user_lactate_threshold_speed: Optional[float] = None
