+++ src/backend/MyAIRunningMate.Python/app/schemas/ingestion/user_metrics_schema.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 class UserMetricsSchema(BaseModel):
     user_volumetric_oxygen_max: Optional[float] = None
-    user_max_heart_rate: Optional[float] = None
+    user_max_heart_rate: Optional[int] = None
     user_lactate_threshold_heart_rate: Optional[float] = None
     user_lactate_threshold_power: Optional[float] = None
     user_lactate_threshold_speed: Optional[float] = None
